Added file acl authentication inspection service including test codes(file.service.test.js)
Added string to array utility including test codes(string.util.test.js)

Deleted authenticated
- authenticated API is no use
