+++ test/string.util.test.js
@@ -1,4 +1,5 @@
 const {isCaseInsensitiveEqual} = require('../util/string-util')
+const {strToArr} = require("../util/file-util");
 
 describe('Util Test', () => {
     test('/string case insensitive equal', () => {
@@ -13,6 +14,22 @@ describe('Util Test', () => {
         for (let t of test) {
             const result = isCaseInsensitiveEqual(t.src, t.target);
             expect(result).toBe(t.result);
+        }
+    });
+});
+
+describe('File String to Array Test', () => {
+    test('/acl authentication test', () => {
+        const test = [
+            {str: "['Open']", result: 1},
+            {str: "['Open'", result: 0},
+            {str: "['Open', 'Closed']", result: 2},
+            {str: undefined, result: 0}
+        ];
+
+        for (let i of test) {
+            const result = strToArr(i.str);
+            expect(result.length).toBe(i.result);
         }
     });
 });
